Baymodel create should allow no network_driver specified
When a kubernetes baymodel is created, the current API validation
only allows an explicit setting of "flannel" for the network-driver
attribute. This validation needs to be changed to allow a user to
create a kubernetes baymodel without explicitly specifying a
network driver. (The same goes for docker swarm and mesos baymodels,
that is, baymodel create should allow the user not to specify
a network-driver, but this is the current behavior).

When no network-driver is selected for baymodel create for a
given COE, Magnum should choose the default network driver for that
COE. Note that this change set only handles the validation for
the case where no network-driver is selected; mapping this to
the COE-dependent default network driver will be done in other
commits (e.g. https://review.openstack.org/#/c/224367/ for swarm).

Note that no new unit test cases were added for this patch since
test_create_baymodel_with_network_driver and
test_create_baymodel_with_no_network_driver test cases in
test_baymodel.py should cover this change.

Change-Id: I0735f79354bc9fdb141e7a559517268e8ab648ec
Closes-Bug: 1501077