Fix skipping of tests
The plugin defines a 'magnum' option in the '[service_available]' group
but does not respect it, so if the plugin is installed but
[service_available]/magnum=false, the tests will still run unless
blacklisted. This patch corrects the issue by defining the tempest
skip_checks class method. This also requires using the correct tempest
base class as recommended in the tempest guide[1].

[1] https://docs.openstack.org/tempest/latest/write_tests.html

Change-Id: I6b9a8f802129fa800d9f4a3eadd8943962c951c0