Fix HTTP 400 error
This is kind of a race condition issue, Tempest is doing a wait and check
when it is trying to delete a cluster. However, at the very end of the cluster
deletion process, a cluster GET request may get a HTTP 400 error which
will cause the Tempest case failed. This HTTP 400 exception reported
by a cluster GET request should be caught and take cluster as still existing.

Task: 40120
Story: 2007833

Change-Id: Ic97abf8d4d584f35b18bff9eab5d3e5ac80c0f02